Fix CA/residue indexing bugs corrupting correlation matrices and network
The correlation/communication descriptors and the PCA network were computed
on the wrong atoms. MINDIST/NONBOND/SALTBRIDGES/HBONDS/INTERACTIONS were
correct; MI, GC, COMMPROP and the derived ADJACENCY/network were not.

Descriptors (topo_traj.py, main.py):
- calphas[i] stored the residue index (atoms_to_resids[...]) instead of the
  CA/C5' atom index, so frame_coords[calphas[i]] read the wrong atom.
- corr_indices was calphas.keys() (0..n-1), so correlation coords sliced the
  first n atoms instead of the CA/C5' atoms. On 1kx5 that meant the first 41
  residues' atoms stood in for all 1268 residues.
- get_chunk_cp divided the already-averaged mean by n_frames again, so COMMPROP
  became ~sum(d^2) instead of the Ca-distance variance.
- calphas / corr_indices are now keyed by residue index via atoms_to_resids, so
  they stay aligned with resids_to_noh even when atom order != residue order
  (non-monotonic resSeq, insertion codes, string-sorted chains).

Network (runner.py, topo_traj.py, generals.py, read_files.py):
- Persist a canonical node->residue mapping (save_atom_mapping) in matrix/
  residue order and load it in process_graphs, instead of re-deriving it from
  the PDB in atom-index order (which could mislabel nodes).

Network robustness (networkparameters.py, generals.py, graph_constructor.py):
- find_alternative_paths used residue numbers as graph node indices; now uses
  the atom_mapping node indices and errors if a residue can't be mapped.
- Yen's k-shortest-paths ordered candidates lexicographically by node index;
  now orders them by path cost via a (cost, tie, path) heap.
- Match graph files by exact name instead of prefix (avoids Graph=1 also
  matching graph_cutoff_10.json).
- Compare min-dist against float(distance_cutoff) instead of int(...).

+++ src/compass/network/graph_constructor.py
@@ -57,7 +57,7 @@ class GraphConstructor:
         # Add edges based on the distance and adjacency matrices
         for i in range(num_nodes):
             for j in range(i + 1, num_nodes):
-                if min_dist_matrix[i, j] < int(distance_cutoff) and \
+                if min_dist_matrix[i, j] < float(distance_cutoff) and \
                         adjacency_matrix[i, j] > 0:
                     G.add_edge(i, j, weight=adjacency_matrix[i, j])
         return G
